+++ settings.js
@@ -398,14 +398,6 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 })
 .addSwitch({
     category: "Chat",
-    configName: "messageOnMusicRuneDrop",
-    title: "Send a party chat message on MUSIC RUNE I drop",
-    description: "Sends a party chat message when a rare item has dropped by you.",
-    subcategory: "Rare Drops",
-    value: true
-})
-.addSwitch({
-    category: "Chat",
     configName: "messageOnSquidPetDrop",
     title: "Send a party chat message on SQUID PET drop",
     description: "Sends a party chat message when a rare item has dropped by you.",
@@ -897,14 +889,6 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 })
 .addSwitch({
     category: "Alerts",
-    configName: "alertOnMusicRuneDrop",
-    title: "Alert on MUSIC RUNE I drop",
-    description: "Shows a title and plays a sound when a rare item has dropped by you or your party members.",
-    subcategory: "Rare Drops",
-    value: true
-})
-.addSwitch({
-    category: "Alerts",
     configName: "alertOnSquidPetDrop",
     title: "Alert on SQUID PET drop",
     description: "Shows a title and plays a sound when a rare item has dropped by you or your party members.",
@@ -1063,7 +1047,15 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Overlays",
     configName: "seaCreaturesHpOverlay",
     title: "Sea creatures HP",
-    description: `Shows an overlay with the HP of nearby Legendary/Mythic sea creatures when they're in lootshare range. Sea creatures with lower HP come first.`,
+    description: `Shows an overlay with the HP of nearby Legendary/Mythic sea creatures when they're in lootshare range.`,
+    subcategory: "Sea creatures HP",
+    value: true
+})
+.addSwitch({
+    category: "Overlays",
+    configName: "seaCreaturesHpOverlay_immunity",
+    title: "Display immunity",
+    description: `Display ~5 seconds immunity indicator for damage reduction period that some sea creature types have.`,
     subcategory: "Sea creatures HP",
     value: true
 })
@@ -1071,7 +1063,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Overlays",
     configName: "seaCreaturesHpOverlay_maxCount",
     title: "Maximum sea creatures HP count",
-    description: "Show maximum N sea creatures nearby (to limit overlay size).",
+    description: "Show maximum N sea creatures nearby (to limit overlay size). Sea creatures with lower HP come first.",
     options: [1, 20],
     value: 5,
     subcategory: "Sea creatures HP"
@@ -1514,6 +1506,13 @@ ${GRAY}Do ${AQUA}/feeshResetProfitTracker${GRAY} to reset.`,
     configName: "showCaughtTrophyFishRaritiesInOdger",
     title: "Caught trophy fish rarities",
     description: "Render caught trophy fish rarities in Odger's Trophy Fishing GUI.",
+    subcategory: "Item icon"
+})
+.addSwitch({
+    category: "Items and storages",
+    configName: "showMobyDuckProgress",
+    title: "Moby-Duck progress",
+    description: "Render percentage of Moby-Duck evolving progress.",
     subcategory: "Item icon"
 })